@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <AssemblyVersion>1.0.0.0</AssemblyVersion>
   </PropertyGroup>
 
@@ -18,20 +18,8 @@
   </PropertyGroup> 
 
   <ItemGroup>
-    <Compile Include="..\shared\Middleware\requestLogger.cs"><Link>shared/Middleware/requestLogger.cs</Link></Compile>
-    <Compile Include="..\shared\Middleware\robotsText.cs"><Link>shared/Middleware/robotsText.cs</Link></Compile>
-    <Compile Include="..\shared\Middleware\version.cs"><Link>shared/Middleware/version.cs</Link></Compile>
-    <Compile Include="..\shared\Model\Actor.cs"><Link>shared/Model/Actor.cs</Link></Compile>
-    <Compile Include="..\shared\Model\Healthz.cs"><Link>shared/Model/Healthz.cs</Link></Compile>
-    <Compile Include="..\shared\Model\Movie.cs"><Link>shared/Model/Movie.cs</Link></Compile>
-    <Compile Include="..\shared\Version.cs"><Link>shared/Version.cs</Link></Compile>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" /> 
-    <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.7" /> 
     <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.3.1" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc2" />
     <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.8.2" />
     <PackageReference Include="Microsoft.Azure.KeyVault" Version="3.0.4" />
     <PackageReference Include="Microsoft.Azure.Services.AppAuthentication" Version="1.3.1" />
